refactor: consolidate duplicated SSRF validation into shared trait (closes #81)
Extract isExternalUrl/isPrivateIp from MiscController and AdminSettingsController
into a shared ValidatesExternalUrls trait. Both controllers now use the trait
instead of maintaining identical private methods.
