+++ app/Http/Concerns/ValidatesExternalUrls.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace App\Http\Concerns;
+
+/**
+ * Shared SSRF protection for controllers that accept external URLs.
+ * Validates that a URL does not target internal/private networks.
+ */
+trait ValidatesExternalUrls
+{
+    protected function isExternalUrl(string $url): bool
+    {
+        if (! preg_match('#^https?://#i', $url)) {
+            return false;
+        }
+
+        $parsed = parse_url($url);
+        $host = $parsed['host'] ?? '';
+        if (empty($host)) {
+            return false;
+        }
+
+        $ips = gethostbynamel($host);
+        if ($ips === false) {
+            return false;
+        }
+
+        foreach ($ips as $ip) {
+            if ($this->isPrivateIp($ip)) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    private function isPrivateIp(string $ip): bool
+    {
+        return ! filter_var(
+            $ip,
+            FILTER_VALIDATE_IP,
+            FILTER_FLAG_NO_PRIV_RANGE | FILTER_FLAG_NO_RES_RANGE
+        );
+    }
+}
